+++ frontend/components/SelectTemplate/Sidebar/Sidebar.tsx
@@ -4,7 +4,7 @@ import {
   IPageType,
   ITemplate,
   readMeGenerator,
-} from "@/api/generated"
+} from "@/openAPI/generated"
 import { useQuery } from "@tanstack/react-query"
 import { useDebounce } from "use-debounce"
